Matrix Synthesi

+++ C Lang/7. Variável indexada/binary_search.c
@@ -2,11 +2,11 @@
 
 int main(void) {
   long v[] = {15, 4, 7, 5, 6, 3, 2, 11, 21, 14, 0};
-  unsigned int length = sizeof(v) / sizeof(long);
+  unsigned long length = sizeof(v) / sizeof(long);
   long key, lower_limit = 0, upper_limit = length - 1;
 
-  for (unsigned long i = 0; i < length; i++)
-    for (unsigned long j = 0; j < length - 1; j++)
+  for (size_t i = 0; i < length; i++)
+    for (size_t j = 0; j < length - 1; j++)
       if (v[i] < v[j]) {
         long aux = v[i];
         v[i] = v[j];
